adds working, tested event and blob store synchronization

+++ memdam/blobstore/localfolder.py
@@ -10,8 +10,10 @@ class Blobstore(memdam.blobstore.api.Blobstore):
     """
 
     def __init__(self, folder):
-        #TODO: really should be an absolute path...
-        self._folder = folder
+        folder = os.path.abspath(folder)
+        while folder[-1] == '/':
+            folder = folder[:-1]
+        self._folder = folder + '/'
 
     def get_url_prefix(self):
         return "file://" + self._folder
@@ -20,14 +22,27 @@ class Blobstore(memdam.blobstore.api.Blobstore):
         path = self._get_path(blob_id, extension)
         _make_folders(path)
         shutil.copyfile(input_path, path)
+        return self.get_url_prefix() + blob_id.hex + "." + extension
 
     def get_data_to_file(self, blob_id, extension, output_path):
         path = self._get_path(blob_id, extension)
+        if not os.path.exists(path):
+            raise memdam.blobstore.api.MissingBlob()
         shutil.copyfile(path, output_path)
 
     def delete(self, blob_id, extension):
         path = self._get_path(blob_id, extension)
-        os.remove(path)
+        try:
+            os.remove(path)
+        except OSError, e:
+            #don't care what happened as long as the file no longer exists
+            if not os.path.exists(path):
+                return
+            raise e
+
+
+    def exists(self, blob_id, extension):
+        return os.path.exists(self._get_path(blob_id, extension))
 
     def _get_path(self, blob_id, extension):
         """
@@ -52,9 +67,11 @@ class Blobstore(memdam.blobstore.api.Blobstore):
 
 def _make_folders(path):
     """Create the subfolders in our directory"""
-    level_two_folder, _ = os.path.split(path)
-    level_one_folder, _ = os.path.split(level_two_folder)
-    if not os.path.exists(level_one_folder):
-        os.mkdir(level_one_folder)
-    if not os.path.exists(level_two_folder):
-        os.mkdir(level_two_folder)
+    folder, _ = os.path.split(path)
+    try:
+        os.makedirs(folder)
+    except OSError, e:
+        #don't care what happened as long as the folder exists
+        if os.path.exists(folder):
+            return
+        raise e
